requirements.txt
celery==5.5.3
Flask==3.1.2
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
python-dotenv==1.1.1
redis==6.4.0
